Restructure files in folders
